refactor: Replace existing channel
